+++ src/index.ts
@@ -1,14 +1,8 @@
-export * from "./lib/DeleteAction/DeleteAction";
 export * from "./lib/Layout/AuthenticationLayout";
+export * from "./lib/Paging/Paging";
 
 // Panel sidebar layout
 export * from "./lib/Layout/PanelSideBarLayout/PanelSideBarLayout";
 export * from "./lib/Layout/PanelSideBarLayout/PanelSideBar/Context/PanelSideBarContext";
 export * from "./lib/Layout/PanelSideBarLayout/PanelSideBar/Definitions/PanelItem";
-
-// Sidebar layout
-export * from "./lib/Layout/SideBarLayout/SideBarLayout";
-export * from "./lib/Layout/SideBarLayout/SideBarLayoutContext";
-export * from "./lib/SideBar/SideBarMenuContext";
-export * from "./lib/Paging/Paging";
-export * from "./lib/SideBar/ISideBarMenuItem";
+export * from "./lib/Layout/PanelSideBarLayout/PanelSideBar/Definitions/PanelLinkRenderer";