+++ src/lib/Layout/PanelSideBarLayout/PanelSideBar/PanelSideBarToggle.tsx
@@ -1,6 +1,8 @@
 import { faAngleLeft, faAngleRight } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { Button, ButtonProps } from "reactstrap";
+import { usePanelSideBarContext } from "./Context/PanelSideBarContext";
+import classNames from "classnames";
 
 interface PanelSideBarToggleProps extends ButtonProps {
   toggled: boolean;
@@ -8,10 +10,20 @@ interface PanelSideBarToggleProps extends ButtonProps {
 
 export const PanelSideBarToggle = (props: PanelSideBarToggleProps) => {
   const { toggled, ...buttonProps } = props;
+  const { theme } = usePanelSideBarContext();
 
   return (
-    <Button {...buttonProps} id="side-nav-toggle" color="primary">
-      <FontAwesomeIcon icon={toggled ? faAngleRight : faAngleLeft} />
+    <Button
+      {...buttonProps}
+      className={classNames(
+        { "side-nav-toggle-dark": theme == "dark" },
+        { "side-nav-toggle-light": theme == "light" },
+        { "side-nav-toggle-blue": theme == "blue" },
+      )}
+      id="side-nav-toggle"
+      color="primary"
+    >
+      <FontAwesomeIcon className="toggler" icon={toggled ? faAngleRight : faAngleLeft} />
     </Button>
   );
 };