+++ src/lib/Layout/PanelSideBarLayout/PanelSideBarLayout.tsx
@@ -1,11 +1,11 @@
 import classNames from "classnames";
-import { PropsWithChildren, ReactNode, useState } from "react";
-import { DropdownMenu, DropdownToggle, Nav, NavItem, UncontrolledDropdown } from "reactstrap";
-import "../../../../styles/Layout/PanelSideBarLayout.scss";
+import { PropsWithChildren, ReactNode } from "react";
+import "../../../../styles/Layout/index.scss";
 import { PanelSideBar } from "./PanelSideBar/PanelSidebar";
 import { PanelSideBarLayoutContent } from "./PanelSideBarLayoutContent";
 import { PanelSideBarToggle } from "./PanelSideBar/PanelSideBarToggle";
-import { usePanelSideBarContext } from "src/lib/Layout/PanelSideBarLayout/PanelSideBar/Context/PanelSideBarContext";
+import { PanelSidebarNavbar } from "./PanelSideBarNavbar";
+import { usePanelSideBarContext } from "./PanelSideBar/Context/PanelSideBarContext";
 
 export interface PanelSideBarLayoutProps extends PropsWithChildren {
   /**
@@ -20,72 +20,63 @@ export interface PanelSideBarLayoutProps extends PropsWithChildren {
    * The collapsible option to choose.
    */
   collapsible?: boolean;
+  /**
+   * The navbar content on the right.
+   */
+  navbarRightItems?: ReactNode[];
+  /**
+   * The navbar content on the left.
+   */
+  navbarLeftItems?: ReactNode[];
+
+  /**
+   * If using the toggle button instead of the side menu adiacent bar.
+   */
+  useToggleButton?: boolean;
+
+  /**
+   * If use the responsive layout when the screen is sm in order to remove the sidebar overlay.
+   */
+  useResponsiveLayout?: boolean;
 }
 
-export const PanelSideBarLayout = (props: PanelSideBarLayoutProps) => {
-  const { brand, children, footer, collapsible = true } = props;
-
-  const [isOpen, setIsOpen] = useState(true);
-  const toggleSidebar = () => setIsOpen((prev) => !prev);
-
-  const [isUserOpen, setIsUserOpen] = useState(false);
-
+export const PanelSideBarLayout = <TPanelItemId extends string, TPanelItem>(props: PanelSideBarLayoutProps) => {
   const {
-    brand: contextBrand,
-    footer: contextFooter,
-    userDropDownMenu,
-    userDropDownMenuToggle,
-    topBarRightCustomItems = [],
-    topBarLeftCustomItems = [],
-    renderFirstItemsLevelAsTiles,
-  } = usePanelSideBarContext();
+    brand,
+    children,
+    navbarLeftItems,
+    navbarRightItems,
+    footer = undefined,
+    collapsible = true,
+    useToggleButton = false,
+    useResponsiveLayout = false,
+  } = props;
 
+  const { isSidebarOpen, toggleSidebar, renderFirstItemsLevelAsTiles } = usePanelSideBarContext<TPanelItemId, TPanelItem>();
+
+  if (useResponsiveLayout && !useToggleButton) {
+    throw new Error("Responsive layout can be used only with toggle button in the navbar!");
+  }
   return (
     <>
-      <nav id="nav-top" className="panel-layout navbar navbar-expand">
-        {/* Navbar Brand */}
-        <div className="navbar-brand">{brand ?? contextBrand}</div>
-        <Nav tag="div" className="navbar-user flex-row justify-content-between" style={{ marginLeft: 35, marginRight: 48 }}>
-          <span className="d-flex flex-row justify-content-start align-items-center">
-            {/*Left Custom Menu*/}
-            {topBarLeftCustomItems?.map((item, index) => (
-              <NavItem tag="div" key={index} className="navbar-custom-item">
-                {item}
-              </NavItem>
-            ))}
-          </span>
-          <span className="d-flex flex-row justify-content-end align-items-center">
-            {/*Right Custom Menu*/}
-            {topBarRightCustomItems?.map((item, index) => (
-              <NavItem tag="div" key={index} className="navbar-custom-item">
-                {item}
-              </NavItem>
-            ))}
-            {/*Navbar user*/}
-            <NavItem tag="div">
-              <UncontrolledDropdown direction="start" isOpen={isUserOpen}>
-                <DropdownToggle nav className="user-dropdown" onClick={() => setIsUserOpen(!isUserOpen)}>
-                  {userDropDownMenuToggle}
-                </DropdownToggle>
-                <div>
-                  <DropdownMenu end>{userDropDownMenu}</DropdownMenu>
-                </div>
-              </UncontrolledDropdown>
-            </NavItem>
-          </span>
-        </Nav>
-      </nav>
-
+      <PanelSidebarNavbar
+        useToggleButton={useToggleButton}
+        brand={brand}
+        navbarRightItems={navbarRightItems}
+        navbarLeftItems={navbarLeftItems}
+      />
       <section
+        id="main-section"
         className={classNames(
-          { toggled: !isOpen },
+          { toggled: !isSidebarOpen },
+          { "responsive-layout": useResponsiveLayout },
           { "section-no-tiles": !renderFirstItemsLevelAsTiles },
           { "section-tiles": renderFirstItemsLevelAsTiles },
         )}
       >
-        <PanelSideBar />
-        {collapsible && <PanelSideBarToggle onClick={toggleSidebar} toggled={!isOpen} />}
-        <PanelSideBarLayoutContent footer={footer ?? contextFooter}>{children}</PanelSideBarLayoutContent>
+        <PanelSideBar<TPanelItemId, TPanelItem> />
+        {collapsible && !useToggleButton && <PanelSideBarToggle onClick={toggleSidebar} toggled={!isSidebarOpen} />}
+        <PanelSideBarLayoutContent footer={footer}>{children}</PanelSideBarLayoutContent>
       </section>
     </>
   );